Add Preferences to Options and remove leftovers from Skeleton building
